meikuangzhinengjiance_heruizhe/rule-backend-flask/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
